requirements.txt
torch
datasets
transformers
Pillow
mlx
einops